update backend for connect to frontend

+++ backend/src/main/java/com/example/bootcamp/repository/ShopRepository.java
@@ -4,11 +4,10 @@ import com.example.bootcamp.entity.ShopsEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import javax.swing.text.html.Option;
 import java.util.Optional;
 
 @Repository
-public interface ShopRepository extends JpaRepository<ShopsEntity, Long> {
+public interface ShopRepository extends JpaRepository<ShopsEntity, Integer> {
     Optional<ShopsEntity> findByShopSlug(String shopSlug);
     Optional<ShopsEntity> findByUserId(Integer userId);
     boolean existsByShopName(String shopName);
